+++ app/javascript/lib/validation/validate-file-content.js
@@ -12,6 +12,14 @@ import { log } from 'lib/metrics-api'
 import { readFileBytes } from './io'
 import ChunkedLineReader from './chunked-line-reader'
 import { PARSEABLE_TYPES } from 'components/upload/upload-utils'
+import {
+  parseDenseMatrixFile, parseFeaturesFile, parseBarcodesFile, parseSparseMatrixFile
+} from './expression-matrices-validation'
+import {
+  getParsedHeaderLines, parseLine, ParseException,
+  validateUniqueCellNamesWithinFile, validateMetadataLabelMatches, validateGroupColumnCounts
+} from './shared-validation'
+
 
 // from lib/assets/metadata_schemas/alexandria_convention_schema.json (which in turn is from scp-ingest-pipeline/schemas)
 export const REQUIRED_CONVENTION_COLUMNS = [
@@ -28,26 +36,6 @@ export const REQUIRED_CONVENTION_COLUMNS = [
   'species__ontology_label'
 ]
 
-/**
- * ParseException can be thrown when we encounter an error that prevents us from parsing the file further
- */
-function ParseException(key, msg) {
-  this.message = msg
-  this.key = key
-}
-
-/**
- * Splits the line on a delimiter, and
- * removes leading and trailing white spaces and quotes from values
- */
-export function parseLine(line, delimiter) {
-  const splitLine = line.split(delimiter)
-  const parsedLine = new Array(parseLine.length)
-  for (let i = 0; i < splitLine.length; i++) {
-    parsedLine[i] = splitLine[i].trim().replaceAll(/^"|"$/g, '')
-  }
-  return parsedLine
-}
 
 /**
  * Verify headers are unique and not empty
@@ -177,245 +165,6 @@ function validateEqualCount(headers, annotTypes) {
   return issues
 }
 
-/**
- * Verify "GENE" is present as the first column in the first row for an Expression Matrix file
- * Todo: Accept files that are R-formatted as well via this ticket: SCP-3971
- */
-function validateGeneInHeader(headers) {
-  const issues = []
-  if (headers[0].toUpperCase() !== 'GENE') {
-    const msg = 'Dense matrices require the first value of the file to be "GENE". ' +
-     `However, the first value for this file currently is "${headers[0]}".`
-    issues.push(['error', 'format:cap:missing-gene-column', msg])
-  }
-
-  return issues
-}
-
-/**
- * Validate the first line in the Sparse Matrix begins with the string '%%MatrixMarket'
- */
-function validateMTXHeaderLine(line, lineNum, dataObj) {
-  const issues = []
-  const mtxHeader = line.slice(0, 14)
-  if (lineNum !== 0) {
-    return issues
-  }
-  if (mtxHeader !== '%%MatrixMarket') {
-    const msg = 'The first line of the file must begin with the string "%%MatrixMarket". ' +
-     `However, the first value for this file currently is "${mtxHeader}".`
-    issues.push(['error', 'format:cap:missing-mtx-value', msg])
-  }
-
-  return issues
-}
-
-/**
- * Validate all values are numbers outside first column cell name
- */
-function validateValuesAreNumeric(line, isLastLine, lineNum, dataObj) {
-  const issues = []
-  dataObj.rowsWithNonNumericValues = dataObj.rowsWithNonNumericValues ? dataObj.rowsWithNonNumericValues : []
-  // skip first column
-  const lineWithoutFirstColumn = line.slice(1)
-
-  if (lineWithoutFirstColumn.some(isNaN)) {
-    dataObj.rowsWithNonNumericValues.push(lineNum)
-  }
-
-  const numBadRows = dataObj.rowsWithNonNumericValues.length
-  if (isLastLine && numBadRows > 0) {
-    const rowText = numBadRows > 1 ? 'rows' : 'row'
-    const maxLinesToReport = 10
-    let notedBadRows = dataObj.rowsWithNonNumericValues.slice(0, maxLinesToReport).join(', ')
-    if (numBadRows - maxLinesToReport > 0) {
-      notedBadRows += ` and ${numBadRows - maxLinesToReport} more rows`
-    }
-
-    const msg = `All values (other than the first column and header row) in a dense matrix file must be numeric. ` +
-    `Please ensure all values in ${rowText}: ${notedBadRows}, are numbers.`
-    issues.push(['error', 'format:invalid-type:not-numeric', msg])
-  }
-
-  return issues
-}
-
-/**
- * Verify sparse matrix column numbers match
- */
-function validateSparseColumnNumber(line, isLastLine, lineNum, dataObj) {
-  const issues = []
-  dataObj.rowsWithWrongColumnNumbers = dataObj.rowsWithWrongColumnNumbers ? dataObj.rowsWithWrongColumnNumbers : []
-  dataObj.correctNumberOfColumns = dataObj.correctNumberOfColumns ? dataObj.correctNumberOfColumns : ''
-
-  // use the first non-comment or header row to determine correct number of columns
-  if (!line.startsWith('%') && !!line.trim()) {
-    dataObj.correctNumberOfColumns = dataObj.correctNumberOfColumns ? dataObj.correctNumberOfColumns : line.split(' ').length
-    if (dataObj.correctNumberOfColumns !== line.split(' ').length) {
-      dataObj.rowsWithWrongColumnNumbers.push(lineNum)
-    }
-  }
-
-  const numBadRows = dataObj.rowsWithWrongColumnNumbers.length
-  if (isLastLine && numBadRows > 0) {
-    const rowText = numBadRows > 1 ? 'rows' : 'row'
-    const maxLinesToReport = 10
-    let notedBadRows = dataObj.rowsWithWrongColumnNumbers.slice(0, maxLinesToReport).join(', ')
-    if (numBadRows - maxLinesToReport > 0) {
-      notedBadRows += ` and ${numBadRows - maxLinesToReport} more rows`
-    }
-
-    const msg = `All rows must have the same number of columns - ` +
-    `please ensure the number of columns for ${rowText}: ${notedBadRows}, ` +
-    `matches the specificed number of columns-per-row.`
-    issues.push(['error', 'format:mismatch-column-number', msg])
-  }
-
-  return issues
-}
-
-/**
- * Verify sparse matrix has no blank lines
- */
-function validateSparseNoBlankLines(line, isLastLine, lineNum, dataObj) {
-  const issues = []
-  dataObj.blankLineRows = dataObj.blankLineRows ? dataObj.blankLineRows : []
-
-  // use the first non-comment or header row to determine correct number of columns
-  if (!line.trim()) {
-    dataObj.blankLineRows.push(lineNum)
-  }
-
-  const numBadRows = dataObj.blankLineRows.length
-  if (isLastLine && numBadRows > 0) {
-    const rowText = numBadRows > 1 ? 'rows' : 'row'
-    const maxLinesToReport = 10
-    let notedBadRows = dataObj.blankLineRows.slice(0, maxLinesToReport).join(', ')
-    if (numBadRows - maxLinesToReport > 0) {
-      notedBadRows += ` and ${numBadRows - maxLinesToReport} more rows`
-    }
-
-    const msg = `Please ensure there are no blank rows in the file. ` +
-    `Remove or replace the following ${rowText}: ${notedBadRows}, `
-    issues.push(['error', 'format:empty-row', msg])
-  }
-
-  return issues
-}
-
-
-/**
- * Verify dense matrix column numbers match header column numbers
- */
-function validateColumnNumber(line, isLastLine, headers, lineNum, dataObj) {
-  const issues = []
-  dataObj.rowsWithIncorrectColumnNumbers = dataObj.rowsWithIncorrectColumnNumbers ? dataObj.rowsWithIncorrectColumnNumbers : []
-  // use the first header row to determine the correct number of columns all rows should have
-  const correctNumberOfColumns = headers[0].length
-
-  if (correctNumberOfColumns !== line.length) {
-    dataObj.rowsWithIncorrectColumnNumbers.push(lineNum)
-  }
-
-  const numBadRows = dataObj.rowsWithIncorrectColumnNumbers.length
-  if (isLastLine && numBadRows > 0) {
-    const rowText = numBadRows > 1 ? 'rows' : 'row'
-    const maxLinesToReport = 10
-    let notedBadRows = dataObj.rowsWithIncorrectColumnNumbers.slice(0, maxLinesToReport).join(', ')
-    if (numBadRows - maxLinesToReport > 0) {
-      notedBadRows += ` and ${numBadRows - maxLinesToReport} more rows`
-    }
-
-    const msg = `All rows must have the same number of columns - ` +
-    `please ensure the number of columns for ${rowText}: ${notedBadRows}, ` +
-    `matches the file-header-specificed number of ${correctNumberOfColumns} columns-per-row.`
-    issues.push(['error', 'format:mismatch-column-number', msg])
-  }
-
-  return issues
-}
-
-/**
- * Verify row values are each unique for a file
- * Per the logic in ingest pipeline check in Feature and Barcode files
- * that the values in each row are unique from all other rows in the file.
- */
-function validateUniqueRowValuesWithinFile(rawLine, isLastLine, dataObj) {
-  const issues = []
-  const line = rawLine.toString().replace(/^"|,|\s|"$/g, '')
-
-  dataObj.rowValues = dataObj.rowValues ? dataObj.rowValues : new Set()
-  dataObj.duplicateRowValues = dataObj.duplicateRowValues ? dataObj.duplicateRowValues : new Set()
-
-  if (!dataObj.rowValues.has(line)) {
-    dataObj.rowValues.add(line)
-  } else {
-    dataObj.duplicateRowValues.add(rawLine)
-  }
-  if (isLastLine && dataObj.duplicateRowValues.size > 0) {
-    const nameTxt = (dataObj.duplicateRowValues.size > 1) ? 'duplicates' : 'duplicate'
-    const dupString = [...dataObj.duplicateRowValues].slice(0, 10).join(', ')
-    const msg = `Row values must be unique within a file. ${dataObj.duplicateRowValues.size} ${nameTxt} found, including: ${dupString}`
-    issues.push(['error', 'duplicate:values-within-file', msg])
-  }
-  return issues
-}
-
-/**
- * Verify cell names are each unique for a file
- * creates and uses 'cellNames' and 'duplicateCellNames' properties on dataObj to track
- * cell names between calls to this function
- */
-function validateUniqueCellNamesWithinFile(line, isLastLine, dataObj) {
-  const issues = []
-
-  dataObj.cellNames = dataObj.cellNames ? dataObj.cellNames : new Set()
-  dataObj.duplicateCellNames = dataObj.duplicateCellNames ? dataObj.duplicateCellNames : new Set()
-  const cell = line[0]
-
-  if (!dataObj.cellNames.has(cell)) {
-    dataObj.cellNames.add(cell)
-  } else {
-    dataObj.duplicateCellNames.add(cell)
-  }
-  if (isLastLine && dataObj.duplicateCellNames.size > 0) {
-    const nameTxt = (dataObj.duplicateCellNames.size > 1) ? 'duplicates' : 'duplicate'
-    const dupString = [...dataObj.duplicateCellNames].slice(0, 10).join(', ')
-    const msg = `Cell names must be unique within a file. ${dataObj.duplicateCellNames.size} ${nameTxt} found, including: ${dupString}`
-    issues.push(['error', 'duplicate:cells-within-file', msg])
-  }
-  return issues
-}
-
-/**
- * Guess whether column delimiter is comma or tab
- *
- * Consider using `papaparse` NPM package once it supports ES modules.
- * Upstream task: https://github.com/mholt/PapaParse/pull/875
- */
-function sniffDelimiter([line1, line2], mimeType) {
-  const delimiters = [',', '\t']
-  let bestDelimiter
-
-  delimiters.forEach(delimiter => {
-    const numFieldsLine1 = line1.split(delimiter).length
-    const numFieldsLine2 = line2.split(delimiter).length
-
-    if (numFieldsLine1 !== 1 && numFieldsLine1 === numFieldsLine2) {
-      bestDelimiter = delimiter
-    }
-  })
-
-  if (typeof bestDelimiter === 'undefined') {
-    if (mimeType === 'text/tab-separated-values') {
-      bestDelimiter = '\t'
-    } else {
-      // fall back on comma -- which may give the most useful error message to the user
-      bestDelimiter = ','
-    }
-  }
-  return bestDelimiter
-}
 
 /**
  * Verify cap format for a cluster or metadata file
@@ -441,21 +190,6 @@ function validateCapFormat([headers, annotTypes]) {
     validateGroupOrNumeric(annotTypes),
     validateEqualCount(headers, annotTypes)
   )
-
-  return issues
-}
-
-/**
- * Verify cap format for an expression matrix file
- *
- * The "cap" for an expression matrix file is the first row
- */
-function validateDenseHeader([headers]) {
-  let issues = []
-  if (!headers) {
-    return [['error', 'format:cap:no-cap-row', 'File does not have a non-empty header row']]
-  }
-  issues = issues.concat(validateGeneInHeader(headers))
 
   return issues
 }
@@ -497,90 +231,6 @@ function validateRequiredMetadataColumns(parsedHeaders) {
   return issues
 }
 
-/**
- * Verify that, for id columns with a corresponding label column, no label is shared across two or more ids.
- * The main circumstance this is aimed at checking is the 'Excel drag error', in which by drag-copying a row, the
- * label is copied correctly, but the id string gets numerically incremented
- */
-function validateMetadataLabelMatches(headers, line, isLastLine, dataObj) {
-  const issues = []
-  const excludedColumns = ['NAME']
-  // if this is the first time through, identify the columns to check, and initialize data structures to track mismatches
-  if (!dataObj.dragCheckColumns) {
-    dataObj.dragCheckColumns = headers[0].map((colName, index) => {
-      const labelColumnIndex = headers[0].indexOf(`${colName }__ontology_label`)
-      if (excludedColumns.includes(colName) ||
-        colName.endsWith('ontology_label') ||
-        headers[1][index] === 'numeric' ||
-        labelColumnIndex === -1) {
-        return null
-      }
-      // for each column, track a hash of label=>value,
-      // and also a set of mismatched values--where the same label is used for different ids
-      return { colName, index, labelColumnIndex, labelValueMap: {}, mismatchedVals: new Set() }
-    }).filter(c => c)
-  }
-  // for each column we need to check, see if there is a corresponding filled-in label,
-  //  and track whether other ids have been assigned to that label too
-  for (let i = 0; i < dataObj.dragCheckColumns.length; i++) {
-    const dcc = dataObj.dragCheckColumns[i]
-    const colValue = line[dcc.index]
-    const label = line[dcc.labelColumnIndex]
-    if (label.length) {
-      if (dcc.labelValueMap[label] && dcc.labelValueMap[label] !== colValue) {
-        dcc.mismatchedVals.add(label)
-      } else {
-        dcc.labelValueMap[label] = colValue
-      }
-    }
-  }
-
-  // only report out errors if this is the last line of the file so that a single, consolidated message can be displayed per column
-  if (isLastLine) {
-    dataObj.dragCheckColumns.forEach(dcc => {
-      if (dcc.mismatchedVals.size > 0) {
-        const labelString = [...dcc.mismatchedVals].slice(0, 10).join(', ')
-        const moreLabelsString = dcc.mismatchedVals.size > 10 ? ` and ${dcc.mismatchedVals.size - 10} others`: ''
-        issues.push(['error', 'content:metadata:mismatched-id-label',
-          `${dcc.colName} has different ID values mapped to the same label.
-          Label(s) with more than one corresponding ID: ${labelString}${moreLabelsString}`])
-      }
-    })
-  }
-  return issues
-}
-
-/** raises a warning if a group column has more than 200 unique values */
-function validateGroupColumnCounts(headers, line, isLastLine, dataObj) {
-  const issues = []
-  const excludedColumns = ['NAME']
-  if (!dataObj.groupCheckColumns) {
-    dataObj.groupCheckColumns = headers[0].map((colName, index) => {
-      if (excludedColumns.includes(colName) || colName.endsWith('ontology_label') || headers[1][index] === 'numeric') {
-        return null
-      }
-      return { colName, index, uniqueVals: new Set() }
-    }).filter(c => c)
-  }
-  for (let i = 0; i < dataObj.groupCheckColumns.length; i++) {
-    const gcc = dataObj.groupCheckColumns[i]
-    const colValue = line[gcc.index]
-    if (colValue) { // don't bother adding empty values
-      gcc.uniqueVals.add(colValue)
-    }
-  }
-
-  if (isLastLine) {
-    dataObj.groupCheckColumns.forEach(gcc => {
-      if (gcc.uniqueVals.size > 200) {
-        issues.push(['warn', 'content:group-col-over-200',
-          `${gcc.colName} has over 200 unique values and so will not be visible in plots -- is this intended?`])
-      }
-    })
-  }
-  return issues
-}
-
 /** Verifies cluster file has X and Y coordinate headers */
 function validateClusterCoordinates(headers) {
   const issues = []
@@ -599,59 +249,6 @@ function validateClusterCoordinates(headers) {
   return issues
 }
 
-/** parse a metadata file, and return an array of issues, along with file parsing info */
-export async function parseBarcodesFile(chunker, mimeType, fileOptions) {
-  let issues = []
-
-  const dataObj = {} // object to track multi-line validation concerns
-  await chunker.iterateLines((rawline, lineNum, isLastLine) => {
-    issues = issues.concat(validateUniqueRowValuesWithinFile(rawline, isLastLine, dataObj))
-  })
-  return { issues }
-}
-
-/** parse a metadata file, and return an array of issues, along with file parsing info */
-export async function parseFeaturesFile(chunker, mimeType, fileOptions) {
-  let issues = []
-
-  const dataObj = {} // object to track multi-line validation concerns
-  await chunker.iterateLines((rawline, lineNum, isLastLine) => {
-    issues = issues.concat(validateUniqueRowValuesWithinFile(rawline, isLastLine, dataObj))
-  })
-  return { issues }
-}
-
-/** parse a dense matrix file */
-export async function parseDenseMatrixFile(chunker, mimeType, fileOptions) {
-  const { headers, delimiter } = await getParsedHeaderLines(chunker, mimeType)
-  let issues = validateDenseHeader(headers)
-
-  const dataObj = {} // object to track multi-line validation concerns
-  await chunker.iterateLines((rawLine, lineNum, isLastLine) => {
-    const line = parseLine(rawLine, delimiter)
-    issues = issues.concat(validateValuesAreNumeric(line, isLastLine, lineNum, dataObj))
-    issues = issues.concat(validateColumnNumber(line, isLastLine, headers, lineNum, dataObj))
-    issues = issues.concat(validateUniqueCellNamesWithinFile(line, isLastLine, dataObj))
-    issues = issues.concat(validateMetadataLabelMatches(headers, line, isLastLine, dataObj))
-    issues = issues.concat(validateGroupColumnCounts(headers, line, isLastLine, dataObj))
-    // add other line-by-line validations here
-  })
-  return { issues, delimiter, numColumns: headers[0].length }
-}
-
-/** parse a sparse matrix file */
-export async function parseSparseMatrixFile(chunker, mimeType, fileOptions) {
-  let issues = []
-  const delimter = '  '
-  const dataObj = {} // object to track multi-line validation concerns
-  await chunker.iterateLines((rawLine, lineNum, isLastLine) => {
-    issues = issues.concat(validateMTXHeaderLine(rawLine, lineNum, dataObj))
-    issues = issues.concat(validateSparseColumnNumber(rawLine, isLastLine, lineNum, dataObj))
-    issues = issues.concat(validateSparseNoBlankLines(rawLine, isLastLine, lineNum, dataObj))
-    // add other line-by-line validations here
-  })
-  return { issues, delimter, numColumns: dataObj.correctNumberOfColumns}
-}
 
 /** parse a metadata file, and return an array of issues, along with file parsing info */
 export async function parseMetadataFile(chunker, mimeType, fileOptions) {
@@ -691,24 +288,6 @@ export async function parseClusterFile(chunker, mimeType) {
   })
 
   return { issues, delimiter, numColumns: headers[0].length }
-}
-
-/**
- * reads in a two lines to be used as header lines, sniffs the delimiter,
- * and returns the lines parsed by the sniffed delimiter
- */
-export async function getParsedHeaderLines(chunker, mimeType) {
-  const headerLines = []
-  await chunker.iterateLines((line, lineNum, isLastLine) => {
-    headerLines.push(line)
-  }, 2)
-  if (headerLines.length < 2 || headerLines.some(hl => !hl)) {
-    throw new ParseException('format:cap:missing-header-lines',
-      `Your file is missing newlines or is missing some of the required header lines`)
-  }
-  const delimiter = sniffDelimiter(headerLines, mimeType)
-  const headers = headerLines.map(l => parseLine(l, delimiter))
-  return { headers, delimiter }
 }
